requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
blas=1.0=mkl
brotlipy=0.7.0=py310h7f8727e_1002
bzip2=1.0.8=h7b6447c_0
ca-certificates=2023.01.10=h06a4308_0
certifi=2022.12.7=py310h06a4308_0
cffi=1.15.1=py310h5eee18b_3
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.0.4=py310h06a4308_0
cryptography=39.0.1=py310h9ce1e76_0
cuda-cudart=11.8.89=0
cuda-cupti=11.8.87=0
cuda-libraries=11.8.0=0
cuda-nvrtc=11.8.89=0
cuda-nvtx=11.8.86=0
cuda-runtime=11.8.0=0
ffmpeg=4.3=hf484d3e_0
filelock=3.9.0=py310h06a4308_0
flit-core=3.8.0=py310h06a4308_0
freetype=2.12.1=h4a9f257_0
giflib=5.2.1=h5eee18b_3
gmp=6.2.1=h295c915_3
gmpy2=2.1.2=py310heeb90bb_0
gnutls=3.6.15=he1e5248_0
huggingface_hub=0.10.1=py310h06a4308_0
idna=3.4=py310h06a4308_0
intel-openmp=2021.4.0=h06a4308_3561
jinja2=3.1.2=py310h06a4308_0
joblib=1.1.1=py310h06a4308_0
jpeg=9e=h5eee18b_1
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libcublas=11.11.3.6=0
libcufft=10.9.0.58=0
libcufile=1.6.0.25=0
libcurand=10.3.2.56=0
libcusolver=11.4.1.48=0
libcusparse=11.7.5.86=0
libdeflate=1.17=h5eee18b_0
libffi=3.4.2=h6a678d5_6
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.16=h7f8727e_2
libidn2=2.3.2=h7f8727e_0
libnpp=11.8.0.86=0
libnvjpeg=11.9.0.86=0
libpng=1.6.39=h5eee18b_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.16.0=h27cfd23_0
libtiff=4.5.0=h6a678d5_2
libunistring=0.9.10=h27cfd23_0
libuuid=1.41.5=h5eee18b_0
libwebp=1.2.4=h11a3e52_1
libwebp-base=1.2.4=h5eee18b_1
lz4-c=1.9.4=h6a678d5_0
markupsafe=2.1.1=py310h7f8727e_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py310h7f8727e_0
mkl_fft=1.3.1=py310hd6ae3a3_0
mkl_random=1.2.2=py310h00e6091_0
mpc=1.1.0=h10f8cd9_1
mpfr=4.0.2=hb69a4c5_1
mpmath=1.2.1=pypi_0
ncurses=6.4=h6a678d5_0
nettle=3.7.3=hbbd107a_1
networkx=2.8.4=py310h06a4308_1
nltk=3.7=pyhd3eb1b0_0
numpy=1.23.5=py310hd5efca6_0
numpy-base=1.23.5=py310h8e6c178_0
openh264=2.1.1=h4ff587b_0
openssl=1.1.1t=h7f8727e_0
packaging=23.0=py310h06a4308_0
pillow=9.4.0=py310h6a678d5_0
pip=23.0.1=py310h06a4308_0
portalocker=2.3.0=py310h06a4308_0
pycparser=2.21=pyhd3eb1b0_0
pyopenssl=23.0.0=py310h06a4308_0
pysocks=1.7.1=py310h06a4308_0
python=3.10.10=h7a1cb2a_2
pytorch=2.0.0=py3.10_cuda11.8_cudnn8.7.0_0
pytorch-cuda=11.8=h7e8668a_3
pytorch-mutex=1.0=cuda
pyyaml=6.0=py310h5eee18b_1
readline=8.2=h5eee18b_0
regex=2022.7.9=py310h5eee18b_0
requests=2.28.1=py310h06a4308_1
setuptools=65.6.3=py310h06a4308_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.41.1=h5eee18b_0
sympy=1.11.1=py310h06a4308_0
tk=8.6.12=h1ccaba5_0
tokenizers=0.11.4=py310h3dcd8bd_1
torchaudio=2.0.0=py310_cu118
torchdata=0.6.0=py310
torchtext=0.15.0=py310
torchtriton=2.0.0=py310
torchvision=0.15.0=py310_cu118
tqdm=4.65.0=py310h2f386ee_0
transformers=4.24.0=py310h06a4308_0
typing-extensions=4.4.0=py310h06a4308_0
typing_extensions=4.4.0=py310h06a4308_0
tzdata=2022g=h04d1e81_0
urllib3=1.26.14=py310h06a4308_0
wheel=0.38.4=py310h06a4308_0
xz=5.2.10=h5eee18b_1
yaml=0.2.5=h7b6447c_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.4=hc292b87_0
